+++ prismmanifest/__init__.py
@@ -1,3 +1,0 @@
-"""PrismManifest package surface used by Prism-Eval."""
-
-__version__ = "0.1.0"